requirements.txt
# Pipecat and related dependencies
pipecat-ai>=0.0.93
# pipecat[deepgram]
python-dotenv>=1.0.0
aiohttp>=3.9.0
loguru>=0.7.0

# Speech-to-Text
deepgram-sdk>=3.0.0

# Text-to-Speech
cartesia>=1.0.0

# LLM
openai>=1.0.0

# Daily.co for WebRTC transport
daily-python>=0.10.0

# VAD
silero-vad>=4.0.0

# Web framework
Flask 
Flask-SocketIO 
Flask-CORS 
python-socketio 
eventlet 
python-dotenv 
requests

websockets>=15.0.1
